factored drop-checking code out into dropck module.
